Deduplicated TableInfo
