+++ discoverx/discovery.py
@@ -13,7 +13,7 @@ logger = logging.Logging()
 class Discovery:
     """ """
 
-    COLUMNS_TABLE_NAME = "system.information_schema.columns"
+    INFORMATION_SCHEMA = "system.information_schema"
     MAX_WORKERS = 10
 
     def __init__(
@@ -36,7 +36,9 @@ class Discovery:
         self._scan_result: Optional[ScanResult] = None
         self.rules: Optional[Rules] = Rules(custom_rules=custom_rules, locale=locale)
 
-    def _msql(self, msql: str, what_if: bool = False, min_score: Optional[float] = None):
+    def _msql(
+        self, msql: str, what_if: bool = False, min_score: Optional[float] = None
+    ):
         logger.debug(f"Executing sql template: {msql}")
 
         msql_builder = Msql(msql)
@@ -73,7 +75,7 @@ class Discovery:
             rule_filter=rules,
             sample_size=sample_size,
             what_if=what_if,
-            columns_table_name=self.COLUMNS_TABLE_NAME,
+            information_schema=self.INFORMAI,
             max_workers=self.MAX_WORKERS,
         )
 
@@ -82,7 +84,9 @@ class Discovery:
 
     def _check_scan_result(self):
         if self._scan_result is None:
-            raise Exception("You first need to scan your lakehouse using Scanner.scan()")
+            raise Exception(
+                "You first need to scan your lakehouse using Scanner.scan()"
+            )
 
     @property
     def scan_result(self):
@@ -136,7 +140,9 @@ class Discovery:
             raise ValueError("search_term has not been provided.")
 
         if not isinstance(search_term, str):
-            raise ValueError(f"The search_term type {type(search_term)} is not valid. Please use a string type.")
+            raise ValueError(
+                f"The search_term type {type(search_term)} is not valid. Please use a string type."
+            )
 
         if by_class is None:
             # Trying to infer the class by the search term
@@ -155,11 +161,15 @@ class Discovery:
                 )
             else:
                 by_class = search_matching_rules[0]
-            logger.friendly(f"Discoverx will search your lakehouse using the class {by_class}")
+            logger.friendly(
+                f"Discoverx will search your lakehouse using the class {by_class}"
+            )
         elif isinstance(by_class, str):
             search_matching_rules = [by_class]
         else:
-            raise ValueError(f"The provided by_class {by_class} must be of string type.")
+            raise ValueError(
+                f"The provided by_class {by_class} must be of string type."
+            )
 
         sql_filter = f"`[{search_matching_rules[0]}]` = '{search_term}'"
         select_statement = (
@@ -209,7 +219,9 @@ class Discovery:
 
         if isinstance(by_classes, str):
             by_classes = [by_classes]
-        elif isinstance(by_classes, list) and all(isinstance(elem, str) for elem in by_classes):
+        elif isinstance(by_classes, list) and all(
+            isinstance(elem, str) for elem in by_classes
+        ):
             by_classes = by_classes
         else:
             raise ValueError(
@@ -229,7 +241,8 @@ class Discovery:
         )
 
         return self._msql(
-            f"SELECT {from_statement}, to_json(struct(*)) AS row_content FROM {from_tables}", min_score=min_score
+            f"SELECT {from_statement}, to_json(struct(*)) AS row_content FROM {from_tables}",
+            min_score=min_score,
         )
 
     def delete_by_class(
@@ -266,7 +279,9 @@ class Discovery:
         Msql.validate_from_components(from_tables)
 
         if (by_class is None) or (not isinstance(by_class, str)):
-            raise ValueError(f"Please provide a class to identify the columns to be matched on the provided values.")
+            raise ValueError(
+                f"Please provide a class to identify the columns to be matched on the provided values."
+            )
 
         if values is None:
             raise ValueError(
@@ -278,7 +293,8 @@ class Discovery:
             value_string = "'" + "', '".join(values) + "'"
         else:
             raise ValueError(
-                f"The provided values {values} have the wrong type. Please provide" f" either a str or List[str]."
+                f"The provided values {values} have the wrong type. Please provide"
+                f" either a str or List[str]."
             )
 
         if not yes_i_am_sure:
@@ -297,4 +313,6 @@ class Discovery:
 
         if delete_result is not None:
             delete_result = delete_result.toPandas()
-            logger.friendlyHTML(f"<p>The affected tables are</p>{delete_result.to_html()}")
+            logger.friendlyHTML(
+                f"<p>The affected tables are</p>{delete_result.to_html()}"
+            )
